requirements.txt
rich>=13.0.0
loguru>=0.7.0
playwright>=1.40.0
pyyaml>=6.0.0
customtkinter>=5.0.0
scipy>=1.9.0
pandas>=1.5.0
plotly>=5.0.0
pytz>=2023.3

